Reorganize module to not expload in warnings when 's3_tests' feature is off
