+++ mountpoint-s3-client/tests/common/creds.rs
@@ -2,24 +2,9 @@
 //!
 //! Some methods are duplicated in `mountpoint-s3::tests::common::creds`.
 
-use aws_config::default_provider::credentials::DefaultCredentialsChain;
-use aws_credential_types::provider::ProvideCredentials;
 use aws_credential_types::Credentials;
-use aws_sdk_sts as sts;
-use aws_sdk_sts::config::Region;
 
-use super::get_test_region;
-
-/// Grab a set of SDK [Credentials] from the default credential provider chain.
-pub async fn get_sdk_default_chain_creds() -> Credentials {
-    let sdk_provider = DefaultCredentialsChain::builder().build().await;
-    let credentials = sdk_provider
-        .provide_credentials()
-        .await
-        .expect("default chain credentials should be available");
-    mask_aws_creds_if_on_gha(&credentials);
-    credentials
-}
+use crate::common::get_test_region;
 
 /// Provide some static credentials that are not valid.
 pub fn get_fake_creds() -> Credentials {
@@ -37,6 +22,35 @@ pub fn get_fake_creds() -> Credentials {
     credentials
 }
 
+/// Detect if running on GitHub Actions (GHA) and if so,
+/// emit masking string to avoid credentials accidentally being printed.
+fn mask_aws_creds_if_on_gha(credentials: &Credentials) {
+    if std::env::var_os("GITHUB_ACTIONS").is_some() {
+        // GitHub Actions aren't aware of these credential strings since we're sourcing them inside the tests.
+        // If we think we're in GitHub Actions environment, register each in stdout.
+        // https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#masking-a-value-in-a-log
+        println!("::add-mask::{}", credentials.access_key_id());
+        println!("::add-mask::{}", credentials.secret_access_key());
+        if let Some(token) = credentials.session_token() {
+            println!("::add-mask::{}", token);
+        }
+    }
+}
+
+/// Grab a set of SDK [Credentials] from the default credential provider chain.
+pub async fn get_sdk_default_chain_creds() -> Credentials {
+    use aws_config::default_provider::credentials::DefaultCredentialsChain;
+    use aws_credential_types::provider::ProvideCredentials;
+
+    let sdk_provider = DefaultCredentialsChain::builder().build().await;
+    let credentials = sdk_provider
+        .provide_credentials()
+        .await
+        .expect("default chain credentials should be available");
+    mask_aws_creds_if_on_gha(&credentials);
+    credentials
+}
+
 /// Takes the provided IAM Policy and assumes the configured subsession role,
 /// applying the given policy to scope down the permissions available.
 ///
@@ -49,7 +63,14 @@ pub fn get_fake_creds() -> Credentials {
 ///
 /// [session_policies]: https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies.html#policies_session
 pub async fn get_scoped_down_credentials<Policy: Into<String>>(policy: Policy) -> Credentials {
-    let sts_client = get_test_sdk_sts_client().await;
+    use aws_sdk_sts as sts;
+
+    let config = aws_config::from_env()
+        .region(sts::config::Region::new(get_test_region()))
+        .load()
+        .await;
+    let sts_client = sts::Client::new(&config);
+
     let assume_role_response = sts_client
         .assume_role()
         .role_arn(get_subsession_iam_role())
@@ -74,30 +95,7 @@ pub async fn get_scoped_down_credentials<Policy: Into<String>>(policy: Policy) -
     credentials
 }
 
-/// Detect if running on GitHub Actions (GHA) and if so,
-/// emit masking string to avoid credentials accidentally being printed.
-fn mask_aws_creds_if_on_gha(credentials: &Credentials) {
-    if std::env::var_os("GITHUB_ACTIONS").is_some() {
-        // GitHub Actions aren't aware of these credential strings since we're sourcing them inside the tests.
-        // If we think we're in GitHub Actions environment, register each in stdout.
-        // https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#masking-a-value-in-a-log
-        println!("::add-mask::{}", credentials.access_key_id());
-        println!("::add-mask::{}", credentials.secret_access_key());
-        if let Some(token) = credentials.session_token() {
-            println!("::add-mask::{}", token);
-        }
-    }
-}
-
 /// ARN of an AWS IAM Role that can be assumed by individual tests to scope down permissions.
 fn get_subsession_iam_role() -> String {
     std::env::var("S3_SUBSESSION_IAM_ROLE").expect("Set S3_SUBSESSION_IAM_ROLE to run integration tests")
-}
-
-async fn get_test_sdk_sts_client() -> sts::Client {
-    let config = aws_config::from_env()
-        .region(Region::new(get_test_region()))
-        .load()
-        .await;
-    sts::Client::new(&config)
 }
